requirements.txt
Flask
Flask-SocketIO
Flask-Cors
requests
openpyxl
pywebview
